Refactor disassemble to a separate file within the evmdis package, instead of main package.
